add index and schema for skills icons
